+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.0.3
+- udjusted code to set popup size according to image size
+- added closing popup window on middle click
+- reduced size of background grid
+- reduced step size of mouse wheel zoom listener
+
 0.0.2
 - fix translations
 - added option to disable fitting window size to image